Django-Cms-Google-Analytics-Poc/requirements.txt
beautifulsoup4==4.4.1
decorator==4.0.4
Django==1.8.4
django-appconf==1.0.1
django-ckeditor==4.5.1
django-classy-tags==0.6.2
django-cms==3.1.3
django-compress==1.0.1
django-compressor==1.5
django-cors-headers==1.1.0
django-reversion==1.8.5
django-sekizai==0.8.2
django-treebeard==3.0
djangocms-admin-style==0.2.8
djangocms-text-ckeditor==2.6.0
html5lib==0.999999
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
lxml==3.5.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wheel==0.24.0
